requirements.txt
Flask>=0.9
redis>=2.6.2
msgpack-python>=0.4.8
